Signal handler: Fix mishandling of faults on 32-on-64, on Linux (Hardy)
Ignore the top 16 bits of REG_CS etc., because some kernel versions
leave these undefined when writing out the signal frame.  This causes
the signal handler to treat trusted faults as untrusted faults, which
causes it to crash when trying to restore %gs.

An alternative fix would be to change the segment register types in
src/trusted/service_runtime/arch/x86_32/nacl_signal_32.h from uint32_t
to uint16_t.  However, there is a comment in that header file saying
that the struct has to match up with src/trusted/gdb_rsp/abi.cc.
Given that the GDB RSP support + debug stub is not covered by
automated tests at the moment (and I don't know how to test it), I am
reluctant to change the struct layout right now.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1486
TEST=run_signal_frame_test: before the fix is applied, this test crashes
  on the 32-bit-on-64-bit Hardy bots

Review URL: http://codereview.chromium.org/6596097

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4438 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: cc69a1f76d8ce64aab2128465b129e78e3306ec8

+++ src/trusted/service_runtime/linux/nacl_signal_32.c
@@ -34,12 +34,25 @@ void NaClSignalContextFromHandler(struct NaClSignalContext *sigCtx,
   sigCtx->edi = mctx->gregs[REG_EDI];
   sigCtx->ebp = mctx->gregs[REG_EBP];
   sigCtx->flags = mctx->gregs[REG_EFL];
-  sigCtx->cs = mctx->gregs[REG_CS];
-  sigCtx->ss = mctx->gregs[REG_SS];
-  sigCtx->ds = mctx->gregs[REG_DS];
-  sigCtx->es = mctx->gregs[REG_ES];
-  sigCtx->fs = mctx->gregs[REG_FS];
-  sigCtx->gs = mctx->gregs[REG_GS];
+   /*
+    * We need to drop the top 16 bits with the casts below.  In some
+    * situations, Linux does not assign to the top 2 bytes of the
+    * REG_CS array entry when writing %cs to the stack (and similarly
+    * for the other segment registers).  Therefore we need to drop the
+    * undefined top 2 bytes.
+    *
+    * This happens in 32-bit processes running on the 64-bit kernel
+    * from Ubuntu Hardy, but not on 32-bit kernels.  The kernel
+    * version in Ubuntu Lucid also does not have this problem.
+    *
+    * See http://code.google.com/p/nativeclient/issues/detail?id=1486
+    */
+  sigCtx->cs = (uint16_t) mctx->gregs[REG_CS];
+  sigCtx->ss = (uint16_t) mctx->gregs[REG_SS];
+  sigCtx->ds = (uint16_t) mctx->gregs[REG_DS];
+  sigCtx->es = (uint16_t) mctx->gregs[REG_ES];
+  sigCtx->fs = (uint16_t) mctx->gregs[REG_FS];
+  sigCtx->gs = (uint16_t) mctx->gregs[REG_GS];
 }
